requirements.txt
streamlit==1.33.0
langchain_community
python-dotenv==1.0.0
openai==1.12.0